Quick update to add a market parameters script as well as configured each parameter script to return a list of ticker symbols/codes.
Added class methods to retrieve bulk data from the stock market, FRED, and the Census Bureau by using the parameter scripts. Each of these methods returns a dictionary of dataframes in the format of -> symbol/code: dataframe

Also removed the code from the main.py script since it is not being used currently. The previous code was used for testing which I've migrated to Google Colab for.